Revert "Be feat/letter crud ming"

+++ back/src/main/java/com/mainproject/back/tag/entity/Tag.java
@@ -21,7 +21,7 @@ public class Tag {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Long tagId;
+  private int tagId;
 
   @Column
   private String name;
